+++ release_notes.md
@@ -73,6 +73,7 @@ First release into griffon artifacts portal.
 1. New mvcPopupButton() node for popup that will display view from another MVCGroup.
 1. New 'linkRenderer' for tableColumnConfig() node for creating a column in JTable that execute an action when clicked.
 1. Add griffon.simplejpa.entityManager.checkThreadSafeLoading configuration key for checking entities lazy loading from thread other than the one that associated with it.
+1. Add griffon.simplejpa.entityManager.defaultFlushMode configuration key for changing EntityManager's flush mode globally.
 1. Add griffon.simplejpa.validation.convertEmptyStringToNull configuration key for automatically converting an empty String into null value when performing validation for a JPA entity.
 1. Minor scaffolding template improvements.
 1. Support grouping in validation (pass grouping class as second argument).